Day_11/tools/pyproject.toml
[project]
name = "tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.14",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
